Fix crash when loading apps list
Some devices have seen crashes when loading the apps list due to an IllegalArgumentException when using the ApplicationInfo.DisplayNameComparator because of a violation of its contract. I don't know the input to cause this issue, but we're already sorting this later in the UI so this prevents a crash and removes useless extra work.