Fix rule database regression
I embarrassingly expected that a hash code would be appropriate for identifying which hosts were contained in the block list. This is not true and can cause very broken behavior even if it worked *mostly*. So I've replaced the MutableIntSet with a MutableScatterSet<String> to still save on memory but also fix the regression.