Do not immediately retry connections if no active network is present
The previous exception that was being thrown from getDnsServers was nonspecific and really would indicate that we aren't connected to an active network. This clears up the purpose of the exception and makes sure that we wait for a new network before trying again instead of looping rapidly and causing OOMs.

Additionally, I added a check to ignore null descriptors that we can get from the system when configuring the VPN and included a specific exception for it. We only want to wait for the network when that happens as well.