More accurately describe expand button action
The arrow icon used for an expandable options item was not really describing what it did by having it point to the side for closed and down for open. It described the state of the menu, but for accessibility reasons, it should be pointing down to show that it can be expanded and up to show that it can be collapsed.