Decouple RuleDatabase from main loop
While not part of this commit, we will be able to load hosts separately from starting the vpn thread so we can save some loading time during thread restarts. The database will just be discarded when the service is killed.

Otherwise, this will make integration testing on the database easier