Move to multi-module with dependency injection
As the app has slowly become more complex with more features, configuration, and UI, I've been running into a few growing pains where business logic has been intertwined with UI logic and a number of embarrassing bugs have made it into production.

In an effort to address this, I have split the project into several modules to force business logic out of UI code and to more clearly define the dependencies of each segment of the app. Alongside that change, I have introduced dependency injection through the hilt library to assist in drawing defined lines between dependencies and their dependents. Not only will this allow for better organization and compile times, but we will also be able to more easily test each component individually.

In order to better facilitate this, I had to change the way that we initialize and access our Preferences and Configuration classes. Preferences is mostly unchanged with an associated SharedPreferences field to allow for unrestricted access to any preference if needed by view models. Configuration was updated to be managed through a new ConfigurationManager so we can make Configuration changes visible to all threads immediately with a similar API to SharedPreferences. Now that manager is just passed around via DI.