Prevent the vpn thread from reloading when stopped by the user
You could run into an ANR if there was no active network and pressed any option to stop. This would be because the thread is already expecting to be interrupted by the network callback when a new network is available and would reload once that signal was sent. However, if you try to stop the thread while it's in this loop, the main thread will hang on waiting for the vpn thread to join.