Only stop vpnThread if it is initialized
In some cases where the VPN has not been prepared and the user tries to start the VPN, it will immediately be stopped. However, at this point, the vpnThread instance will not have been initialized and will crash the service. This also can happen if the system destroys the service before it has the chance to start and the service crashes once again. Here we just check before we try accessing it.