Replace polling with mio
Between learning more about the cargo ecosystem with the amount of support that goes into mio and the lead maintainer of polling stepping away due to personal reasons, this felt like the right move. mio was basically a drop-in replacement and we don't have to use a bunch of unsafe blocks anymore.