Real test for IPv6
The previous IPv6 tester made an assumption that your network supported IPv6 if you had IPv6 DNS servers. While it's not a bad guess, you can more accurately test this by attempting to send a packet to an IPv6 address and returning true if it doesn't return an error (this checks if the network is unreachable but we don't need to be specific).

This eliminates the need for an IPv6 toggle because we can just autodetect the right option.