Fix destroy crash in RuleDatabaseManager
Fixes race condition where the rule database is destroyed, the service is still active, and the database is prompted to reload at the same time. This makes sure that we ignore any calls to reload once the database is destroyed.

Additionally, this moves from atomic bools to semaphores for blocking more than one database reload.