SingleWriterMultipleReaderFile: Fix file existence checks
Writes previously would erroneously "fail" when making checks for successful deletion of the working file in finishWrite and successful renaming of the working file in failWrite. Renaming can "fail" when creating the active file for the first time. This is solved by checking if the active file exists, and only failing if the rename "failed," the active file didn't already exist and it doesn't exist now. In failWrite, there was a missing check to see if the working file existed already before attempting to delete it. It would always fail if the working file did not already exist.